tests: Extract ContainerDisk code to containerdisk.go
As part of re-organizing the tests/utils codebase, the container disk
related helpers are moved to a dedicated file.
